+++ src/parser/api/cpp/input_parser.h
@@ -24,7 +24,7 @@
 #include <memory>
 
 #include "parser/api/cpp/command.h"
-#include "parser/flex_parser.h"
+#include "parser/parser.h"
 
 namespace cvc5 {
 namespace parser {
@@ -128,14 +128,12 @@ class CVC5_EXPORT InputParser
   std::unique_ptr<SymbolManager> d_allocSm;
   /** Symbol manager */
   SymbolManager* d_sm;
-  /** whether to use flex */
-  bool d_useFlex;
   /** Incremental string input language */
   std::string d_istringLang;
   /** Incremental string name */
   std::string d_istringName;
   /** The parser */
-  std::unique_ptr<FlexParser> d_fparser;
+  std::unique_ptr<Parser> d_fparser;
 };
 
 }  // namespace parser