+++ src/parser/smt2/smt2_cmd_parser.cpp
@@ -22,7 +22,7 @@
 namespace cvc5 {
 namespace parser {
 
-Smt2CmdParser::Smt2CmdParser(Smt2LexerNew& lex,
+Smt2CmdParser::Smt2CmdParser(Smt2Lexer& lex,
                              Smt2State& state,
                              Smt2TermParser& tparser)
     : d_lex(lex), d_state(state), d_tparser(tparser)
@@ -284,8 +284,7 @@ std::unique_ptr<Command> Smt2CmdParser::parseNextCommand()
       }
       else
       {
-        Term func = d_state.getSolver()->mkConst(t, name);
-        cmd.reset(new DeclareFunctionCommand(name, func, t));
+        cmd.reset(new DeclareFunctionCommand(name, t));
       }
     }
     break;
@@ -332,8 +331,7 @@ std::unique_ptr<Command> Smt2CmdParser::parseNextCommand()
       Sort t = d_tparser.parseSort();
       std::vector<Term> terms = d_tparser.parseTermList();
       Trace("parser") << "declare pool: '" << name << "'" << std::endl;
-      Term pool = d_state.getSolver()->declarePool(name, t, terms);
-      cmd.reset(new DeclarePoolCommand(name, pool, t, terms));
+      cmd.reset(new DeclarePoolCommand(name, t, terms));
     }
     break;
     // (declare-sort <symbol> <numeral>)
@@ -857,16 +855,8 @@ std::unique_ptr<Command> Smt2CmdParser::parseNextCommand()
       Grammar* g = d_tparser.parseGrammarOrNull(sygusVars, name);
 
       Trace("parser-sygus") << "Define synth fun : " << name << std::endl;
-      Solver* slv = d_state.getSolver();
-      Term fun =
-          isInv ? (g == nullptr ? slv->synthInv(name, sygusVars)
-                                : slv->synthInv(name, sygusVars, *g))
-                : (g == nullptr ? slv->synthFun(name, sygusVars, range)
-                                : slv->synthFun(name, sygusVars, range, *g));
-
-      Trace("parser-sygus") << "...read synth fun " << name << std::endl;
       d_state.popScope();
-      cmd.reset(new SynthFunCommand(name, fun, sygusVars, range, isInv, g));
+      cmd.reset(new SynthFunCommand(name, sygusVars, range, g));
     }
     break;
     case Token::EOF_TOK: