+++ test/unit/api/java/TermTest.java
@@ -1046,6 +1046,36 @@ class TermTest
   }
 
   @Test
+  void getRealAlgebraicNumber() throws CVC5ApiException
+  {
+    d_solver.setOption("produce-models", "true");
+    d_solver.setLogic("QF_NRA");
+    Sort realsort = d_solver.getRealSort();
+    Term x = d_solver.mkConst(realsort, "x");
+    Term x2 = d_solver.mkTerm(MULT, x, x);
+    Term two = d_solver.mkReal(2, 1);
+    Term eq = d_solver.mkTerm(EQUAL, x2, two);
+    d_solver.assertFormula(eq);
+    // Note that check-sat should only return "sat" if libpoly is enabled.
+    // Otherwise, we do not test the following functionality.
+    if (d_solver.checkSat().isSat())
+    {
+      // We find a model for (x*x = 2), where x should be a real algebraic number.
+      // We assert that its defining polynomial is non-null and its lower and
+      // upper bounds are real.
+      Term vx = d_solver.getValue(x);
+      assertTrue(vx.isRealAlgebraicNumber());
+      Term y = d_solver.mkVar(realsort, "y");
+      Term poly = vx.getRealAlgebraicNumberDefiningPolynomial(y);
+      assertTrue(!poly.isNull());
+      Term lb = vx.getRealAlgebraicNumberLowerBound();
+      assertTrue(lb.isRealValue());
+      Term ub = vx.getRealAlgebraicNumberUpperBound();
+      assertTrue(ub.isRealValue());
+    }
+  }
+
+  @Test
   void substitute()
   {
     Term x = d_solver.mkConst(d_solver.getIntegerSort(), "x");