+++ NEWS.md
@@ -17,7 +17,16 @@ This file contains a summary of important user-visible changes.
        `--print-cores-full`. Setting this option to true will print all
        assertions in the unsat core, regardless of whether they are named. This
        option also impacts how timeout cores are printed.
+- API: Added support for querying the values of real algebraic numbers in the
+       API. In particular, the methods `Term::isRealAlgebraicNumber()`,
+       `Term::getRealAlgebraicNumberDefiningPolynomial()`,
+       `Term::getRealAlgebraicNumberLowerBound()`, and
+       `Term::getRealAlgebraicNumberUpperBound()` may now be used to query
+       the contents of terms corresponding to real algebraic numbers.
 - Removed support for the ANTLR parser and parsing for the TPTP language.
+- API: Removed API support for the deprecated SyGuS 2.0 command
+       `Solver::synthInv`. This method is equivalent to `Solver::synthFun`
+       with a Boolean range sort.
 
 cvc5 1.0.5
 ==========